Remove Authorization header on redirects.
App is changing to use v4 of the github upload-artifact hook. In both v3 and v4, the github download
link redirects from a github.com link to some other host that contains authentication information in
the URL itself. But the host/webserver that's being redirected to changed.

Golang's http client preserves `Authorization` headers on redirect (in this case, the contents of
the header is the github API token). This seems to deviate from "the standard" (i.e: what cURL and
Chrome) do.

The webserver hosting the files for the v3 upload script ignored this unnecessary Authorization
header (or maybe accepted it as it was a more obviously github associated service). But the host for
the v4 files returns a 400 noting a bad Authorization header.

So this commit changes the behavior to always remove the Authorization header on a redirect (maybe
it should be preserved if the prior request + current request have a matching `Host`?).

It is noted and accepted that this code is incorrect if the http client accessed with concurrent
application level request. This is due to the split up nature of `CheckRedirect` and calling the
`Transport`.

More details on https://github.com/viamrobotics/app/pull/4588